+++ R/bind-dimensions-n.R
@@ -16,13 +16,13 @@ bind_dimensions_n <- function(...) {
 bind_dimensions_n.mcmcarray <- function(...) {
   x <- list(...)
 
-  if(!length(x)) error("one or more objects must be passed to bind_dimensions_n")
+  if(!length(x)) err("one or more objects must be passed to bind_dimensions_n")
 
   if(length(x) == 1) return(x[[1]])
 
   pdims <- lapply(x, pdims)
   if(!all(vapply(pdims, identical, TRUE, pdims[[1]])))
-    error("all objects must have the same parameter dimensions")
+    err("all objects must have the same parameter dimensions")
 
   along <- pdims(x[[1]])
   along <- length(along) + 1L
@@ -37,13 +37,13 @@ bind_dimensions_n.mcmcarray <- function(...) {
 bind_dimensions_n.mcmcr <- function(...) {
   x <- list(...)
 
-  if(!length(x)) error("one or more objects must be passed to bind_dimensions_n")
+  if(!length(x)) err("one or more objects must be passed to bind_dimensions_n")
 
   if(length(x) == 1) return(x[[1]])
 
   pdims <- lapply(x, pdims)
   if(!all(vapply(pdims, identical, TRUE, pdims[[1]])))
-    error("all objects must have the same parameter dimensions")
+    err("all objects must have the same parameter dimensions")
 
   along <- pdims(x[[1]])
   along <- vapply(along, length, 1L) + 1L