+++ R/bind-iterations.R
@@ -15,13 +15,13 @@ bind_iterations <- function(x, x2, ...) {
 #' @describeIn bind_iterations Binds two mcarray objects by their iterations
 #' @export
 bind_iterations.mcarray <- function(x, x2, ...) {
-  if (!is.mcarray(x2)) error("x2 must be an mcarray")
+  if (!is.mcarray(x2)) err("x2 must be an mcarray")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   x <- abind::abind(x, x2, along = ndims(x)-1)
   set_class(x, "mcarray")
@@ -30,19 +30,19 @@ bind_iterations.mcarray <- function(x, x2, ...) {
 #' @describeIn bind_iterations Binds two mcmc objects by their iterations
 #' @export
 bind_iterations.mcmc <- function(x, x2, ...) {
-  if (!is.mcmc(x2)) error("x2 must be an mcmc")
+  if (!is.mcmc(x2)) err("x2 must be an mcmc")
 
   x <- sort(x)
   x2 <- sort(x2)
 
   if (!identical(parameters(x), parameters(x2)))
-    error("x and x2 must have the same parameters")
+    err("x and x2 must have the same parameters")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   x <- abind::abind(x, x2, along = 1)
   as.mcmc(x)
@@ -51,19 +51,19 @@ bind_iterations.mcmc <- function(x, x2, ...) {
 #' @describeIn bind_iterations Binds two mcmc.list objects by their iterations
 #' @export
 bind_iterations.mcmc.list <- function(x, x2, ...) {
-  if (!(is.mcmc.list(x2) || is.mcmc(x2))) error("x2 must be an mcmc.list")
+  if (!(is.mcmc.list(x2) || is.mcmc(x2))) err("x2 must be an mcmc.list")
 
   x <- sort(x)
   x2 <- sort(x2)
 
   if (!identical(parameters(x), parameters(x2)))
-    error("x and x2 must have the same parameters")
+    err("x and x2 must have the same parameters")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   x <- mapply(x, x2, FUN = bind_iterations, SIMPLIFY = FALSE)
   set_class(x, "mcmc.list")
@@ -73,13 +73,13 @@ bind_iterations.mcmc.list <- function(x, x2, ...) {
 #' @export
 bind_iterations.mcmcarray <- function(x, x2, ...) {
 
-  if (!is.mcmcarray(x2)) error("x2 must be an mcmcarray")
+  if (!is.mcmcarray(x2)) err("x2 must be an mcmcarray")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   x <- abind::abind(x, x2, along = 2)
   set_class(x, "mcmcarray")
@@ -88,19 +88,19 @@ bind_iterations.mcmcarray <- function(x, x2, ...) {
 #' @describeIn bind_iterations Binds two mcmcr objects by their iterations
 #' @export
 bind_iterations.mcmcr <- function(x, x2, ...) {
-  if (!is.mcmcr(x2)) error("x2 must be an mcmcr")
+  if (!is.mcmcr(x2)) err("x2 must be an mcmcr")
 
   x <- sort(x)
   x2 <- sort(x2)
 
   if (!identical(parameters(x), parameters(x2)))
-    error("x and x2 must have the same parameters")
+    err("x and x2 must have the same parameters")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   x <- mapply(x, x2, FUN = bind_iterations, SIMPLIFY = FALSE)
   set_class(x, "mcmcr")