+++ R/bind-parameters.R
@@ -16,13 +16,13 @@ bind_parameters <- function(x, ...) {
 #' @describeIn bind_parameters Binds two mcmc objects by their parameters
 #' @export
 bind_parameters.mcmc <- function(x, x2, ...) {
-  if (!is.mcmc(x2)) error("x2 must be an mcmc.list")
+  if (!is.mcmc(x2)) err("x2 must be an mcmc.list")
 
   if (length(intersect(parameters(x), parameters(x2))))
-    error("x and x2 must not have any of the same parameters")
+    err("x and x2 must not have any of the same parameters")
 
   if (!identical(niters(x), niters(x2)))
-    error("x and x2 must have the same number of iterations")
+    err("x and x2 must have the same number of iterations")
 
   x <- abind::abind(x, x2, along = 2)
   x <- as.mcmc(x)
@@ -32,16 +32,16 @@ bind_parameters.mcmc <- function(x, x2, ...) {
 #' @describeIn bind_parameters Binds two mcmc.list objects by their parameters
 #' @export
 bind_parameters.mcmc.list <- function(x, x2, ...) {
-  if (!is.mcmc.list(x2)) error("x2 must be an mcmc.list")
+  if (!is.mcmc.list(x2)) err("x2 must be an mcmc.list")
 
   if (length(intersect(parameters(x), parameters(x2))))
-    error("x and x2 must not have any of the same parameters")
+    err("x and x2 must not have any of the same parameters")
 
     if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   if (!identical(niters(x), niters(x2)))
-    error("x and x2 must have the same number of iterations")
+    err("x and x2 must have the same number of iterations")
 
   x <- mapply(x, x2, FUN = bind_parameters, SIMPLIFY = FALSE)
   set_class(x, "mcmc.list")
@@ -50,16 +50,16 @@ bind_parameters.mcmc.list <- function(x, x2, ...) {
 #' @describeIn bind_parameters Binds two mcmcr objects by their parameters
 #' @export
 bind_parameters.mcmcr <- function(x, x2, ...) {
-  if (!is.mcmcr(x)) error("x2 must be an mcmcr")
+  if (!is.mcmcr(x)) err("x2 must be an mcmcr")
 
   if (length(intersect(parameters(x), parameters(x2))))
-    error("x and x2 must not have any of the same parameters")
+    err("x and x2 must not have any of the same parameters")
 
     if (!identical(nchains(x), nchains(x2)))
-    error("x and x2 must have the same number of chains")
+    err("x and x2 must have the same number of chains")
 
   if (!identical(niters(x), niters(x2)))
-    error("x and x2 must have the same number of iterations")
+    err("x and x2 must have the same number of iterations")
 
   x <- c(x, x2)
   x <- set_class(x, "mcmcr")