+++ R/combine-samples.R
@@ -20,25 +20,25 @@ combine_samples <- function(x, x2, fun = mean, ...) {
 #' @describeIn combine_samples Combine samples of two mcmcarray objects
 #' @export
 combine_samples.mcmcarray <- function(x, x2, fun = mean, ...) {
-  if (!is.mcmcarray(x2)) error("x2 must be an mcmcarray")
+  if (!is.mcmcarray(x2)) err("x2 must be an mcmcarray")
   check_function(fun)
 
   x <- bind_dimensions(x, x2)
   x <- apply(x, (2:ndims(x) - 1L), fun)
 
-  if(!identical(dims(x), dims(x2))) error("function fun must return a scalar")
+  if(!identical(dims(x), dims(x2))) err("function fun must return a scalar")
   set_class(x, "mcmcarray")
 }
 
 #' @describeIn combine_samples Combine samples of two mcmcr objects
 #' @export
 combine_samples.mcmcr <- function(x, x2, fun = mean, ...) {
-  if (!is.mcmcr(x2)) error("x2 must be an mcmcr")
+  if (!is.mcmcr(x2)) err("x2 must be an mcmcr")
   x <- sort(x)
   x2 <- sort(x2)
 
   if (!identical(pdims(x), pdims(x2)))
-    error("x and x2 must have the same parameter dimensions")
+    err("x and x2 must have the same parameter dimensions")
 
   x <- bind_dimensions(x, x2)
   x <- lapply(x, function(x, fun) {apply(x, (2:ndims(x) - 1L), fun)}, fun = fun)
@@ -46,6 +46,6 @@ combine_samples.mcmcr <- function(x, x2, fun = mean, ...) {
   x <- set_class(x, "mcmcr")
 
   if (!identical(pdims(x), pdims(x2)))
-    error("function fun must return a scalar")
+    err("function fun must return a scalar")
   x
 }