+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import coda checkr
+#' @import coda err checkr
 #' @importFrom graphics plot
 #' @importFrom stats coef terms
 NULL